OSIS-850 Move cancel_proposal in delete file
